requirements.txt
# erg.ai - AI-Powered Rowing Performance Analytics
# Python 3.9+ required

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
numpy==1.26.4

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2
scipy==1.11.4

# Optional: For better performance
# numexpr==2.8.8
# bottleneck==1.3.7echo pyyaml==6.0.3

pyyaml==6.0.3
